Track a per host last audited time instead of global.
This allows us to more gracefully and automatically handle the scenario
where host(s) do not submit logs before the audit report runs. In that
case, we'd either miss their logs, or need to manually re-run with a past
start filter time to see these logs. With this update, we automatically
pull logs since the last we saw for the host.

This does away with the need for a state file for logaudit all together. I
also remove the global filter time flag as it no longer makes a lot of
sense.

I also update error messages to not start with a capital nor end with a
period.